[WIP] inner image of magnifier moves, not magnifying effect yet and weird padding
